refactor: decouple generateAccessMetaData from note
