+++ packages/note-access/.eslintrc.js
@@ -8,4 +8,7 @@ module.exports = {
         node: true,
         jest: true,
     },
+    rules: {
+        'no-console': 'off',
+    }
 };
