+++ packages/note-access/package.json
@@ -5,8 +5,8 @@
   "scripts": {
     "build": "yarn clean && babel --copy-files --out-dir ./lib ./src",
     "clean": "shx rm -rf ./lib",
-    "lint": "eslint --ignore-path ../../.eslintignore .",
-    "test": "mocha ./test --bail --colors --exit --recursive --reporter spec --timeout 0 --trace-warnings",
+    "lint": "eslint --ignore-path  ../../.eslintignore .",
+    "test": "jest",
     "watch": "yarn build --watch"
   },
   "repository": {
@@ -41,7 +41,15 @@
   "devDependencies": {
     "@babel/plugin-proposal-class-properties": "^7.5.5",
     "@babel/plugin-transform-runtime": "^7.6.2",
+    "babel-eslint": "^10.0.3",
     "babel-plugin-module-resolver": "^3.2.0",
-    "got": "^9.6.0"
+    "got": "^9.6.0",
+    "web3-utils": "^1.2.4"
+  },
+  "dependencies": {
+    "@aztec/secp256k1": "^1.0.2",
+    "@trust/webcrypto": "^0.9.2",
+    "jest": "^24.9.0",
+    "web3": "^1.2.4"
   }
 }
