+++ packages/note-access/src/index.js
@@ -1,5 +1,3 @@
-import addAccess from './metadata/addAccess';
-import METADATA_AZTEC_DATA_LENGTH from './config/constants';
-import encryptedViewingKey from './encryptedViewingKey';
+import grantAccess from './grantAccess';
 
-export { addAccess, encryptedViewingKey, METADATA_AZTEC_DATA_LENGTH };
+export default grantAccess;
